+++ src/app/mypage/consumer/wishlist/components/tabFilter.tsx
@@ -19,7 +19,7 @@ export default function TabFilter({
   separator = "|",
 }: TabFilterProps) {
   return (
-    <div className="flex items-center pl-4 gap-3 text-md font-medium mb-12.5 h-9">
+    <div className="flex items-center  gap-3 text-md font-medium mb-12.5 h-9">
       {items.map((item, index) => (
         <div key={item.id} className="flex items-center gap-3">
           <button
